requirements.txt
Pillow==8.2.0
tweepy==3.8.0
APScheduler==3.7.0
selenium==3.141.0
pytwitcasting==1.0.0
chromedriver-binary==92.0.4515.107.0
